agents/requirements.txt
fastapi
uvicorn
python-dotenv
requests
google-genai
backboard-sdk
